Updating Pest grammar with examples. Update build script.

+++ supported_grammars/build.sh
@@ -1,17 +1,29 @@
 #!/bin/sh
 
-#for i in abnf antlr2 antlr3 antlr4 bison cocor iso14977 javacc jison lark lbnf pegen pegjs pest rex sablecc w3cebnf xtext
-# iso14977
-for i in javacc jison lark lbnf pegen pegjs pest rex sablecc w3cebnf xtext
+cwd=`pwd`
+for i in abnf antlr2 antlr3 antlr4 bison cocor iso14977 javacc jison lark lbnf pegen pegjs pest rex sablecc w3cebnf xtext
 do
-    cd $i
-	rm -rf Generated
+    echo $i
+    cd $cwd/$i
+    rm -rf Generated
     if [ -f "pom.xml" ]
     then
-        trgen --template-sources-directory ../templates
-		dotnet build Generated/*.csproj
+        trgen --template-sources-directory ../templates > /dev/null 2>&1
+	if [ $? != 0 ]
+	then
+	    echo "Failed trgen"
+            rm -rf Generated
+	    continue;
+	fi
+	dotnet build Generated/*.csproj > /dev/null 2>&1
+	if [ $? != 0 ]
+	then
+	    echo "Failed build"
+            rm -rf Generated
+	    continue;
+	fi
     else
         echo "No pom.xml in $i"
     fi
-    cd ..
+    rm -rf Generated
 done
